refactored seed
